EMS: RestorePageMap function

+++ src/Spice86.Core/Emulator/InterruptHandlers/Dos/Ems/EmmMapping.cs
@@ -4,16 +4,16 @@ public record EmmMapping {
     public const int Size = 4;
     private byte[] _data { get; init; } = new byte[4];
     
-    public int Handle {
-        get => (_data[0] & 0xFF) | ((_data[1] & 0xFF) << 8);
+    public ushort Handle {
+        get => (ushort) ((_data[0] & 0xFF) | ((_data[1] & 0xFF) << 8));
         set {
             _data[0] = (byte) (value & 0xFF);
             _data[1] = (byte) ((value >> 8) & 0xFF);
         }
     }
 
-    public int Page {
-        get => (_data[2] & 0xFF) | ((_data[3] & 0xFF) << 8);
+    public ushort Page {
+        get => (ushort) ((_data[2] & 0xFF) | ((_data[3] & 0xFF) << 8));
         set {
             _data[2] = (byte) (value & 0xFF);
             _data[3] = (byte) ((value >> 8) & 0xFF);
